+++ NEWS.md
@@ -1,6 +1,6 @@
 # miniPCH 0.4.0
 
-* added S3 methods summary, print, plot and autoplot
+* added S3 methods summary, print, plot and autoplot for miniPCH class
 
 # miniPCH 0.3.2